relocate the data struct namespace to /schema (out of the megadumper addon ns)
