+++ addons/megadumper/formatters/interface.go
@@ -1,6 +1,6 @@
 package formatters
 
-import "github.com/robbyt/llm_proxy/addons/megadumper/schema"
+import "github.com/robbyt/llm_proxy/schema"
 
 // MegaDumpFormatter abstracts the different types of log storage formats
 type MegaDumpFormatter interface {
