+++ addons/megadumper/formatters/json.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/robbyt/llm_proxy/addons/megadumper/schema"
+	"github.com/robbyt/llm_proxy/schema"
 )
 
 type JSON struct {
