+++ addons/stdoutWriter/stdoutwriter.go
@@ -1,36 +1,16 @@
 package stdoutWriter
 
 import (
-	"encoding/json"
-
 	px "github.com/kardianos/mitmproxy/proxy"
+	"github.com/robbyt/llm_proxy/schema"
 	log "github.com/sirupsen/logrus"
 )
 
 const UnknownAddr = "unknown"
 
-type LogLine struct {
-	ClientAddress string `json:"client_address"`
-	Method        string `json:"method"`
-	URL           string `json:"url"`
-	StatusCode    int    `json:"status_code"`
-	ContentLength int    `json:"content_length"`
-	Duration      int64  `json:"duration_ms"`
-	ContentType   string `json:"content_type,omitempty"`
-	XreqID        string `json:"x_request_id,omitempty"`
-	ProxyID       string `json:"proxy_id,omitempty"`
-}
-
-func (obj *LogLine) ToJSONstr() string {
-	jsonData, err := json.Marshal(obj)
-	if err != nil {
-		log.Errorf("Failed to marshal object to JSON: %v", err)
-	}
-	return string(jsonData)
-}
-
 func getClientAddr(f *px.Flow) string {
 	if f == nil || f.ConnContext == nil || f.ConnContext.ClientConn == nil || f.ConnContext.ClientConn.Conn == nil {
+		// Ugh != nil
 		return UnknownAddr
 	}
 	remote := f.ConnContext.ClientConn.Conn.RemoteAddr()
@@ -40,13 +20,13 @@ func getClientAddr(f *px.Flow) string {
 	return remote.String()
 }
 
-func NewLogLine(f *px.Flow, doneAt int64) *LogLine {
+func NewLogLine(f *px.Flow, doneAt int64) *schema.ConnectionStatsLogger {
 	if f == nil {
 		log.Error("Flow object is nil")
 		return nil
 	}
 
-	logOutput := &LogLine{
+	logOutput := &schema.ConnectionStatsLogger{
 		ClientAddress: getClientAddr(f),
 		Method:        f.Request.Method,
 		URL:           f.Request.URL.String(),
