+++ schema/utils/CanPrint.go
